[STRUCTURE] [REFACTO] Rename source folders to kebab-case
Rename camelCase/PascalCase folders to kebab-case (component files keep their
PascalCase names):
- editors/MapEditor → map-editor, SoundEditor → sound-editor,
  SpriteEditor → sprite-editor
- shared/authOverlay → auth-overlay, canvas/gameCanvas → game-canvas,
  luaEnvManager → lua-env-manager, snackBar → snack-bar

Update all import references. Document the folder-naming rule in AGENTS.md
(kebab-case folders; PascalCase component files; camelCase hooks/utils).

Verified: typecheck, lint, jest (7/7), and build all pass.
