+++ src/modules/create/game-editor/editors/map-editor/SpritePicker.tsx
