+++ src/modules/create/game-editor/editors/sound-editor/InstrumentEditor.tsx
