+++ src/modules/create/game-editor/editors/sound-editor/Note.ts
