+++ src/modules/create/game-editor/editors/sound-editor/components/ControlButtons.tsx
