+++ src/modules/create/game-editor/editors/sound-editor/constants/instruments.ts
