+++ src/modules/create/game-editor/editors/sound-editor/instruments/contrabass.json
