+++ src/modules/create/game-editor/editors/sound-editor/instruments/guitar.json
