+++ src/modules/create/game-editor/editors/sound-editor/styles/ProgressBar.styles.ts
