+++ src/modules/create/game-editor/editors/sound-editor/styles/SoundEditor.styles.ts
