+++ src/modules/create/game-editor/editors/sprite-editor/Color.tsx
