+++ src/modules/hub/components/GameViewer.tsx
@@ -10,7 +10,7 @@ import {
 import { GameProvider, ProviderEventType } from "@providers/GameProvider";
 import { useUser } from "@providers/UserProvider";
 import { type SpriteRendererHandle } from "@shared/canvas/RendererHandle";
-import { type EnvData } from "@shared/luaEnvManager/LuaEnvironmentManager";
+import { type EnvData } from "@shared/lua-env-manager/LuaEnvironmentManager";
 import * as urls from "@shared/navigation/routes";
 import { LocalStorageManager } from "@utils/LocalStorageManager";
 import { getCachedProjectImageUrl } from "@utils/projectImageCache";
