+++ src/shared/auth-overlay/AuthOverlay.styles.ts
