+++ src/shared/auth-overlay/AuthOverlay.tsx
