+++ src/shared/canvas/game-canvas/GameCanvas.tsx
@@ -1,9 +1,9 @@
 import { SoundProvider } from "@providers/editors/SoundProvider";
 import { MusicPlayer } from "@shared/audio/MusicPlayer";
 import { CanvasProps, StyledCanvas } from "@shared/canvas/Canvas";
-import { KeyHandler } from "@shared/canvas/gameCanvas/KeyHandler";
+import { KeyHandler } from "@shared/canvas/game-canvas/KeyHandler";
 import { SpriteRendererHandle } from "@shared/canvas/RendererHandle";
-import { EnvData, LuaEnvironmentManager } from "@shared/luaEnvManager/LuaEnvironmentManager";
+import { EnvData, LuaEnvironmentManager } from "@shared/lua-env-manager/LuaEnvironmentManager";
 
 import React, { forwardRef, useCallback, useEffect, useImperativeHandle, useRef } from "react";
