+++ src/shared/navbar/login/Login.tsx
@@ -1,4 +1,4 @@
-import AuthOverlay from "@shared/authOverlay/AuthOverlay";
+import AuthOverlay from "@shared/auth-overlay/AuthOverlay";
 import { NavElem } from "@shared/navbar/NavElem";
 
 import { FC, useCallback, useEffect, useState } from "react";
